Graphql restructuring completed
