+++ packages/headless/.env.example
@@ -1,0 +1,18 @@
+NODE_ENV = DEVELOPMENT
+
+#Database to use
+DB = MONGODB
+
+# HTTP layer to use
+API = GRAPHQL
+
+# MONGODB
+MONGODB_URI = mongodb://localhost:27017/bs-commerce-dev
+
+# MYSQL
+MYSQL_DB_HOST = localhost
+MYSQL_DB_PORT = 3306
+MYSQL_DB_USER = root
+MYSQL_DB_PASS = ''
+MYSQL_DB_DIALECT = mysql
+MYSQL_DB_NAME_TEST = bs-commerce-dev
