+++ packages/headless/src/internal/graphql/graphql.module.resolver.ts
@@ -1,0 +1,9 @@
+import { GraphqlModule } from './graphql.init';
+import { WishListModule as WishListGraphqlModule } from '../../modules/wishlist/wishlist.graphql.module';
+
+export const ResolveGraphqlModule = () => {
+  return [
+    WishListGraphqlModule,
+    ...GraphqlModule()
+  ];
+};
