+++ packages/headless/src/internal/rest/rest.module.resolver.ts
@@ -1,0 +1,7 @@
+import { WishListModule as WishListRestModule } from '../../modules/wishlist/wishlist.rest.module';
+
+export const ResolveRestModule = () => {
+  return [
+    WishListRestModule,
+  ];
+};
