--- packages/headless/src/rest.module.dependency.ts
@@ -1,7 +1,0 @@
-import { WishListModule as WishListRestModule } from './modules/wishlist/wishlist.rest.module';
-
-export const ResolveRestModule = () => {
-  return [
-    WishListRestModule,
-  ];
-};
